content/canvas/test/test_2d.shadow.blur.low.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.shadow.blur.low</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#ff0';
ctx.fillRect(0, 0, 100, 50);
ctx.shadowColor = '#00f';
ctx.shadowOffsetY = 25;
for (var x = 0; x < 100; ++x) {
    ctx.save();
    ctx.beginPath();
    ctx.rect(x, 0, 1, 50);
    ctx.clip();
    ctx.shadowBlur = x;
    ctx.fillRect(-200, -200, 500, 200);
    ctx.restore();
}

todo(false, "test completed successfully"); // (Bug 483989)
SimpleTest.finish();

});
</script>
